Installed & tested react-native-rename to change app name & bundle
