+++ android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.reactnativeboilerplate",
+    package = "com.paceprogrammer.rnboilerplate",
 )
 
 android_resource(
     name = "res",
-    package = "com.reactnativeboilerplate",
+    package = "com.paceprogrammer.rnboilerplate",
     res = "src/main/res",
 )
