+++ android/app/src/debug/java/com/paceprogrammer/rnboilerplate/ReactNativeFlipper.java
@@ -4,7 +4,7 @@
  * <p>This source code is licensed under the MIT license found in the LICENSE file in the root
  * directory of this source tree.
  */
-package com.reactnativeboilerplate;
+package com.paceprogrammer.rnboilerplate;
 
 import android.content.Context;
 import com.facebook.flipper.android.AndroidFlipperClient;
