+++ android/app/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="com.reactnativeboilerplate">
+  package="com.paceprogrammer.rnboilerplate">
 
     <uses-permission android:name="android.permission.INTERNET" />
