+++ android/app/src/main/java/com/paceprogrammer/rnboilerplate/MainActivity.java
@@ -1,4 +1,4 @@
-package com.reactnativeboilerplate;
+package com.paceprogrammer.rnboilerplate;
 
 import com.facebook.react.ReactActivity;
 import com.facebook.react.ReactActivityDelegate;
@@ -14,7 +14,7 @@ public class MainActivity extends ReactActivity {
    */
   @Override
   protected String getMainComponentName() {
-    return "ReactNativeBoilerplate";
+    return "RNB";
   }
 
   /**
