+++ android/app/src/main/java/com/paceprogrammer/rnboilerplate/MainApplication.java
@@ -1,4 +1,4 @@
-package com.reactnativeboilerplate;
+package com.paceprogrammer.rnboilerplate;
 
 import android.app.Application;
 import android.content.Context;
@@ -9,7 +9,7 @@ import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.config.ReactFeatureFlags;
 import com.facebook.soloader.SoLoader;
-import com.reactnativeboilerplate.newarchitecture.MainApplicationReactNativeHost;
+import com.paceprogrammer.rnboilerplate.newarchitecture.MainApplicationReactNativeHost;
 import java.lang.reflect.InvocationTargetException;
 import java.util.List;
 
@@ -73,7 +73,7 @@ public class MainApplication extends Application implements ReactApplication {
          We use reflection here to pick up the class that initializes Flipper,
         since Flipper library is not available in release mode
         */
-        Class<?> aClass = Class.forName("com.reactnativeboilerplate.ReactNativeFlipper");
+        Class<?> aClass = Class.forName("com.paceprogrammer.rnboilerplate.ReactNativeFlipper");
         aClass
             .getMethod("initializeFlipper", Context.class, ReactInstanceManager.class)
             .invoke(null, context, reactInstanceManager);
