+++ android/app/src/main/java/com/paceprogrammer/rnboilerplate/newarchitecture/MainApplicationReactNativeHost.java
@@ -1,4 +1,4 @@
-package com.reactnativeboilerplate.newarchitecture;
+package com.paceprogrammer.rnboilerplate.newarchitecture;
 
 import android.app.Application;
 import androidx.annotation.NonNull;
@@ -19,9 +19,9 @@ import com.facebook.react.fabric.CoreComponentsRegistry;
 import com.facebook.react.fabric.FabricJSIModuleProvider;
 import com.facebook.react.fabric.ReactNativeConfig;
 import com.facebook.react.uimanager.ViewManagerRegistry;
-import com.reactnativeboilerplate.BuildConfig;
-import com.reactnativeboilerplate.newarchitecture.components.MainComponentsRegistry;
-import com.reactnativeboilerplate.newarchitecture.modules.MainApplicationTurboModuleManagerDelegate;
+import com.paceprogrammer.rnboilerplate.BuildConfig;
+import com.paceprogrammer.rnboilerplate.newarchitecture.components.MainComponentsRegistry;
+import com.paceprogrammer.rnboilerplate.newarchitecture.modules.MainApplicationTurboModuleManagerDelegate;
 import java.util.ArrayList;
 import java.util.List;
