+++ ios/RNB/AppDelegate.mm
@@ -44,7 +44,7 @@ static NSString *const kRNConcurrentRoot = @"concurrentRoot";
 #endif
 
   NSDictionary *initProps = [self prepareInitialProps];
-  UIView *rootView = RCTAppSetupDefaultRootView(bridge, @"ReactNativeBoilerplate", initProps);
+  UIView *rootView = RCTAppSetupDefaultRootView(bridge, @"RNB", initProps);
 
   if (@available(iOS 13.0, *)) {
     rootView.backgroundColor = [UIColor systemBackgroundColor];
