+++ ios/RNB/Images.xcassets/AppIcon.appiconset/120.png
