+++ ios/RNB/Images.xcassets/AppIcon.appiconset/48.png
