+++ ios/RNB/Images.xcassets/AppIcon.appiconset/50.png
