+++ ios/RNB/Images.xcassets/AppIcon.appiconset/60.png
