+++ ios/RNB/Images.xcassets/AppIcon.appiconset/76.png
