+++ ios/RNB/Images.xcassets/AppIcon.appiconset/80.png
