+++ ios/RNB/Images.xcassets/AppIcon.appiconset/Contents.json
