+++ package.json
@@ -63,6 +63,8 @@
     "jest": "^26.6.3",
     "metro-react-native-babel-preset": "^0.70.3",
     "react-dom": "18.0.0",
+    "react-native-rename": "^2.9.0",
+    "react-native-version": "^4.0.0",
     "react-test-renderer": "18.0.0",
     "typescript": "^4.4.4"
   },
